+++ src/wp/jbuilder-impl/wp.ml
@@ -1,3 +1,0 @@
-type t = Dune | Jbuilder
-let t = Jbuilder
-let dune2 = false