+++ test/blackbox-tests/test-cases/odig-files-workspace/run.t
@@ -1,6 +1,6 @@
 Odig files should be detected relative to the package directory
   $ dune build @install --root .
-  $ cat project/foobar.install
+  $ cat _build/default/project/foobar.install
   lib: [
     "_build/install/default/lib/foobar/META"
     "_build/install/default/lib/foobar/dune-package"